+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This document contains all changes to the crate since version 0.1.8.
 
+## 0.3.0 (unreleased)
+
+### Breaking changes
+
+- Made all quadrature rules store their nodes and weights sorted in ascending order by node.
+ This means that all functions that return some view of the nodes (and weights) now return them in this sorted order.
+ This affects the Gauss-Legendre, Gauss-Hermite and Gauss-Chebyshev rules.
+ The affected functions are `QuadratureRule::iter()`, `QuadratureRule::into_iter()`, `QuadratureRule::nodes()`, `QuadratureRule::weights()`, `QuadratureRule::as_node_weight_pairs()` and `QuadratureRule::into_node_weight_pairs()`.
+- Made the `Simpson` and `Midpoint` rules not allocate any memory. This removes all the functions on those types that access some view of the nodes.
+- Made the constructors of `Simpson` and `Midpoint` into `const` functions.
+- Made the `new` functions that can only fail for a single reason return an `Option` instead of a `Result<Self, CustomError>`.
+
 ## 0.2.3
 
 - Make the `QuadratureRule::integrate` functions take a `FnMut` instead of a `Fn`.